+++ packages/phone_log/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.7+1
+Removed a `dart:async` import that isn't required for \>=Dart 2.1.
+Require \>=Dart 2.1.
+
+## 0.0.7
+Migrate to Android v2 embedding.
+
 ## 0.0.6
 Added a fallback value for 'number' field in case CACHED_MATCHED_NUMBER is empty.